move compiler-related functions into their own file
